+++ default.nix
@@ -1,8 +1,9 @@
 { mkDerivation, aeson, base, binary, bits, cereal, cereal-vector
-, containers, criterion, deepseq, DPutils, hashable, lens
-, log-domain, mtl, OrderedBits, primitive, QuickCheck, smallcheck
-, stdenv, tasty, tasty-quickcheck, tasty-smallcheck, tasty-th, text
-, vector, vector-binary-instances, vector-th-unbox
+, containers, criterion, deepseq, DPutils, hashable, hashtables
+, lens, log-domain, mtl, OrderedBits, primitive, QuickCheck
+, smallcheck, stdenv, tasty, tasty-quickcheck, tasty-smallcheck
+, tasty-th, text, unordered-containers, vector, vector-algorithms
+, vector-binary-instances, vector-th-unbox
 }:
 mkDerivation {
   pname = "PrimitiveArray";
@@ -10,21 +11,22 @@ mkDerivation {
   src = ./.;
   libraryHaskellDepends = [
     aeson base binary bits cereal cereal-vector containers deepseq
-    DPutils hashable lens log-domain mtl OrderedBits primitive
-    QuickCheck smallcheck text vector vector-binary-instances
-    vector-th-unbox
+    DPutils hashable hashtables lens log-domain mtl OrderedBits
+    primitive QuickCheck smallcheck text unordered-containers vector
+    vector-algorithms vector-binary-instances vector-th-unbox
   ];
   testHaskellDepends = [
     aeson base binary bits cereal cereal-vector containers deepseq
-    DPutils hashable lens log-domain mtl OrderedBits primitive
-    QuickCheck smallcheck tasty tasty-quickcheck tasty-smallcheck
-    tasty-th text vector vector-binary-instances vector-th-unbox
+    DPutils hashable hashtables lens log-domain mtl OrderedBits
+    primitive QuickCheck smallcheck tasty tasty-quickcheck
+    tasty-smallcheck tasty-th text unordered-containers vector
+    vector-algorithms vector-binary-instances vector-th-unbox
   ];
   benchmarkHaskellDepends = [
     aeson base binary bits cereal cereal-vector containers criterion
-    deepseq DPutils hashable lens log-domain mtl OrderedBits primitive
-    QuickCheck smallcheck text vector vector-binary-instances
-    vector-th-unbox
+    deepseq DPutils hashable hashtables lens log-domain mtl OrderedBits
+    primitive QuickCheck smallcheck text unordered-containers vector
+    vector-algorithms vector-binary-instances vector-th-unbox
   ];
   homepage = "https://github.com/choener/PrimitiveArray";
   description = "Efficient multidimensional arrays";