+++ lib/Data/PrimitiveArray.hs
@@ -1,13 +1,11 @@
 
-module Data.PrimitiveArray 
+module Data.PrimitiveArray
   ( module Data.PrimitiveArray.Class
   , module Data.PrimitiveArray.Dense
---  , module Data.PrimitiveArray.FillTables
   , module Data.PrimitiveArray.Index
   ) where
 
 import Data.PrimitiveArray.Class
 import Data.PrimitiveArray.Dense
---import Data.PrimitiveArray.FillTables
 import Data.PrimitiveArray.Index