+++ changelog.md
@@ -1,3 +1,9 @@
+0.11.0.0
+--------
+
+- introduction of @Data.PrimitiveArray.Hashed@ which uses the @hashtables@ package and provides
+  sparsified tables.
+
 0.10.0.0
 --------